Refactor: improve cohesion
