+++ web/app/omattiedot/header/HeaderName.jsx
@@ -1,0 +1,21 @@
+import React from 'react'
+import {modelData, modelTitle} from '../../editor/EditorModel'
+import {ISO2FinnishDate} from '../../date/date'
+import Text from '../../i18n/Text'
+
+export const HeaderName = ({henkilö}) => {
+  const nimi = <p>{`${modelData(henkilö, 'etunimet')} ${modelData(henkilö, 'sukunimi')}`}</p>
+  const syntymäaika = modelTitle(henkilö, 'syntymäaika') &&
+    <p className='syntynyt'>
+      <Text name='syntynyt'/>
+      <span> {ISO2FinnishDate(modelTitle(henkilö, 'syntymäaika'))}</span>
+    </p>
+
+
+  return (
+    <div className='header__name'>
+      {nimi}
+      {syntymäaika}
+    </div>
+  )
+}
